refactor(editable): move editable component into interactables directory
