+++ components/views/chat/chatbar/Chatbar.html
@@ -16,7 +16,7 @@
     >
       <terminal-icon size="1x" />
     </div> -->
-    <ChatbarEditable
+    <InteractablesEditable
       v-model="text"
       ref="editable"
       :placeholder="$t('ui.talk')"
